+++ NEWBIE/TO-DO todo-app-main/js/tareas/creación_tareas_/contenedor-tareas.js
@@ -1,0 +1,6 @@
+import { tareasRestantes } from "../../filtros/tareas_restantes.js";
+export const tareas = [];
+
+export const obtenerTareas = () => {
+  return tareas;
+};
